Move post functions to controller
